refactor: Split 16 oversized Go files by concern into sibling files
Sixteen non-test Go files sat in the 600-800 LOC band, approaching the
800-line cap enforced by scripts/verify-service-conventions.sh and
tests/architecture/size_test.go. Each is split into cohesive sibling
files within the same package by moving top-level declarations verbatim,
preserving behavior and all external imports.

Splits (original -> largest resulting file):
- current-account persistence/repository.go: extract entity mapping
- current-account domain/account.go: extract CurrentAccountBuilder
- financial-accounting persistence/repository.go: extract conversions
- financial-accounting client/starlark.go: extract posting handlers
- tenant worker/provisioning_worker.go: extract retry/backoff
- tenant provisioner/postgres_provisioner.go: extract lifecycle ops
- reference-data node/postgres_repository.go: extract traversal/bulk
- mcp-server tools/audit.go: extract reconciliation tools
- mcp-server tools/gateway.go: extract status formatters/converters
- control-plane applier/executor.go: extract convert helpers
- forecasting starlark/runner.go: extract conversion helpers
- market-information persistence/observation_repository.go: extract query
- position-keeping app/container.go: extract shutdown methods
- saga linter.go: extract lintVisitor
- position-tool cmd/import.go: extract instrument-checker registry
- position-tool checkpoint/manager.go: extract scan/domain helpers

All resulting files are now under 700 lines. No escape-hatch markers were
needed; every file decomposed naturally by concern.
